Use Go to build test binary before executing tests
As expected, the OMGFruit directory that this task executes
in did not yet have the `test` binary needed to run the tests.
This commit uses the `go` command to build the binary first.

I'm expecting this to work because the machine that I'm using as the
OSX worker (my own in this case) already has Golang installed. This
would certainly fail on a machine that does not have Go installed.